remove 2 from names
