+++ sample/source/scenes/CMakeLists.txt
@@ -2,6 +2,6 @@ target_sources(${NC_SAMPLE_EXE}
     PRIVATE
         Benchmarks.cpp
         GraphicsTest.cpp
-        PhysicsTest2.cpp
+        PhysicsTest.cpp
         SmokeTest.cpp
 )
