+++ source/engine/physics2/CMakeLists.txt
@@ -2,7 +2,7 @@ target_sources(${NC_ENGINE_LIB}
     PRIVATE
         Constraints.cpp
         EventDispatch.cpp
-        NcPhysicsImpl2.cpp
+        NcPhysicsImpl.cpp
         PhysicsUtility.cpp
         RigidBody.cpp
         Shape.cpp
